Updates to do the following: 1. Remove unused references. 2. Remove unused usings. 3. Remove the MessageWrapper class in the ProtoSerializer. 4. Remove the Registrar class and create a static create method on the Serializer.
