+++ src/Chinchilla.Serializers.Protobuf.Specifications/Chinchilla.Serializers.Protobuf.Specifications.csproj
@@ -48,14 +48,12 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\CommonAssemblyInfo.cs">
+      <Link>CommonAssemblyInfo.cs</Link>
+    </Compile>
     <Compile Include="AddressBook.cs" />
     <Compile Include="ProtobufMessageSerializerSpecification.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
