+++ src/Chinchilla.Serializers.Protobuf/Chinchilla.Serializers.Protobuf.csproj
@@ -36,12 +36,7 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\CommonAssemblyInfo.cs">
@@ -49,7 +44,6 @@
     </Compile>
     <Compile Include="ProtobufMessageSerializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ProtobufRegistrar.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Chinchilla\Chinchilla.csproj">
